Fix the Customize-page-widget-esi-settings-can-not-update bug

+++ src/admin-settings.cls.php
@@ -326,14 +326,16 @@ class Admin_Settings extends Base {
 		$instance[ Conf::OPTION_NAME ][ ESI::WIDGET_O_TTL ] = $ttl ;
 
 		$current = ! empty( $old_instance[ Conf::OPTION_NAME ] ) ? $old_instance[ Conf::OPTION_NAME ] : false ;
-		if ( ! $current || $esi != $current[ ESI::WIDGET_O_ESIENABLE ] ) {
-			Purge::purge_all( 'Wdiget ESI_enable changed' ) ;
-		}
-		elseif ( $ttl != 0 && $ttl != $current[ ESI::WIDGET_O_TTL ] ) {
-			Purge::add( Tag::TYPE_WIDGET . $widget->id ) ;
-		}
+		if ( ! strpos( $_SERVER[ 'HTTP_REFERER' ], '/wp-admin/customize.php') ) {
+			if ( ! $current || $esi != $current[ ESI::WIDGET_O_ESIENABLE ] ) {
+				Purge::purge_all( 'Wdiget ESI_enable changed' ) ;
+			}
+			elseif ( $ttl != 0 && $ttl != $current[ ESI::WIDGET_O_TTL ] ) {
+				Purge::add( Tag::TYPE_WIDGET . $widget->id ) ;
+			}
 
-		Purge::purge_all( 'Wdiget saved' ) ;
+			Purge::purge_all( 'Wdiget saved' ) ;
+		}
 		return $instance ;
 	}
 }
